rename: AaveCustomProvider StarlayProviderManager
